fix: split sample apps into multiple files
